// fe/src/lib/api/generated/model/geographicDistributionSection.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { GeographicDistributionRow } from './geographicDistributionRow';

export interface GeographicDistributionSection {
  denominator: number;
  key: string;
  label: string;
  rows: GeographicDistributionRow[];
}